feat : #13 도서 검색 결과 view 구현

+++ src/main/java/com/t3t/frontserver/book/adaptor/BookCategoryAdaptor.java
@@ -13,5 +13,5 @@ import org.springframework.web.bind.annotation.RequestParam;
 @FeignClient(name = "bookCategoryAdaptor", url = "http://localhost:8081")
 public interface BookCategoryAdaptor {
     @GetMapping(value = "/category/{categoryId}/books")
-    ResponseEntity<BaseResponse<PageResponse<BookSearchResultResponse>>> getBooksByCategoryId(@PathVariable Long categoryId, @RequestParam int pageNo);
+    ResponseEntity<BaseResponse<PageResponse<BookSearchResultResponse>>> getBooksByCategoryId(@PathVariable Integer categoryId, @RequestParam int pageNo);
 }
